requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
APScheduler==3.10.4

# Face Recognition dependencies (Phase 5)
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.1.0
numpy>=1.26.0,<2

# Background validation (YOLOv8 + reference image similarity for attendance)
ultralytics>=8.0.0,<9

